% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{AA}
\alias{AA}
\title{Azimuth Angle}
\usage{
AA(latitude, EA, DA, LST, HRA)
}
\arguments{
\item{latitude}{Location's latitude in degree}

\item{EA}{Elevation Angle in degree}

\item{DA}{Declination Angle in degree}

\item{LST}{Local solar time in hours}

\item{HRA}{Hour Angle in degree}
}
\value{
Azimuth Angle in degree
}
\description{
The azimuth angle is the compass direction from which the sunlight is coming.
}
\examples{
AA(51.53,30,30,1,1)
}

