% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{AM}
\alias{AM}
\title{Air Mass}
\usage{
AM(ZA)
}
\arguments{
\item{ZA}{Sun zenith angle in degree}
}
\value{
Air mass
}
\description{
The proportion of atmosphere that the light must pass through before striking the Earth
}
\examples{
AM(30)
}

