% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{DA}
\alias{DA}
\title{Declination Angle}
\usage{
DA(day)
}
\arguments{
\item{day}{Number of days since the start of the year}
}
\value{
Declination Angle in degree
}
\description{
The declination of the sun is the angle between the equator and
a line drawn from the centre of the Earth to the centre of the sun.
}
\examples{
DA(256)
}

