% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{EOT}
\alias{EOT}
\title{Equation of Time}
\usage{
EOT(day)
}
\arguments{
\item{day}{Number of days since the start of the year}
}
\value{
Equation of Time in minutes
}
\description{
The equation of time (EoT) (in minutes) is an empirical equation that corrects for 
the eccentricity of the Earth's orbit and the Earth's axial tilt.
}
\examples{
EOT(256)
}

