% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{findOptimalTiltAngleYear}
\alias{findOptimalTiltAngleYear}
\title{Optimal Angles Calculation}
\usage{
findOptimalTiltAngleYear(year, latitude, longtitude, GMT, seaLevel)
}
\arguments{
\item{year}{Year number}

\item{latitude}{Location's latitude in degree}

\item{longtitude}{Location's longtitude in degree}

\item{GMT}{Location's GMT offset}

\item{seaLevel}{Location height above sea level in kilometers}
}
\value{
Dataframe contains Solar Energy output from 112 combination of Module Tilt Angle and Module Azimuth Angle
}
\description{
Calculate optimal angles (Module Tilt Angle and Module Azimuth Angle)
to get maximum Solar Energy output for one year.
}
\examples{
df <- findOptimalTiltAngleYear(2015,0,0,0,0)
optimalAzi  <- df$azi[df$solar == max(df$solar)][1]
optimalTilt <- df$tilt[df$solar == max(df$solar)][1]

}

