% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{HRA}
\alias{HRA}
\title{Hour Angle}
\usage{
HRA(LST)
}
\arguments{
\item{LST}{Local solar time in hours}
}
\value{
Hour Angle (HRA) in degree
}
\description{
The Hour Angle converts the local solar time (LST) into 
the number of degrees which the sun moves across the sky.
In the morning the hour angle is negative, in the afternoon the hour angle is positive.
}
\examples{
HRA(1.112565)
}

