% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{is.leapyear}
\alias{is.leapyear}
\title{Check Leap Year}
\usage{
is.leapyear(year)
}
\arguments{
\item{year}{Year number}
}
\value{
TRUE if year is leap year, FALSE if not
}
\description{
Check if year is leap year or not.
}
\examples{
is.leapyear(2016)
}

