% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{LSTM}
\alias{LSTM}
\title{Local Standard Time Meridian}
\usage{
LSTM(delta)
}
\arguments{
\item{delta}{The difference of the Local Time (LT) from Greenwich Mean Time (GMT) in hours}
}
\value{
Local Standard Time Meridian (LSTM) in degree
}
\description{
The difference of the Local Time (LT) from Greenwich Mean Time (GMT) in degree.
}
\examples{
LSTM(10)
}

