% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{S_m}
\alias{S_m}
\title{Solar Radiation on a Tilted Surface}
\usage{
S_m(S_d, EA, AA, tilt, M_AA)
}
\arguments{
\item{S_d}{Direct beam solar radiation intensity in kW/m^2}

\item{EA}{Sun elevation angle in degree}

\item{AA}{Sun azimuth angle in degree}

\item{tilt}{Module tilt angle in degree}

\item{M_AA}{Module azimuth angle in degree}
}
\value{
Solar energy in kW/m^2
}
\description{
Solar energy on a surface with arbitrary tilt and orientation.
}
\examples{
S_m(0.8178,53.0242,94.7078,34,0)
}

