% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{solarOneDay}
\alias{solarOneDay}
\title{Solar Energy For One Day}
\usage{
solarOneDay(year, month, day, latitude, longtitude, GMT, titlAngle, moduleAzi,
  seaLevel, useBin = TRUE)
}
\arguments{
\item{year}{Year number}

\item{month}{Month number}

\item{day}{Day number}

\item{latitude}{Location's latitude in degree}

\item{longtitude}{Location's longtitude in degree}

\item{GMT}{Location's GMT offset}

\item{titlAngle}{Tilt Angle of the surface in degree}

\item{moduleAzi}{Direction the surface is facing}

\item{seaLevel}{Location height above sea level in kilometers}

\item{useBin}{Print solar energy for every hour in day if TRUE}
}
\value{
Combined vector of Solar Energy for one day (kWh/m^2/day) and every hour (kWh/m^2) in that day
}
\description{
Calculate Solar Energy for One day.
}
\examples{
solarOneDay(2015,6,22,51.56,15.32,0,30,180,0.145)
}

