% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{solarOneYear}
\alias{solarOneYear}
\title{Solar Energy For One Year}
\usage{
solarOneYear(year, latitude, longtitude, GMT, titlAngle, moduleAzi, seaLevel,
  useBin = TRUE)
}
\arguments{
\item{year}{Year number}

\item{latitude}{Location's latitude in degree}

\item{longtitude}{Location's longtitude in degree}

\item{GMT}{Location's GMT offset}

\item{titlAngle}{Tilt Angle of the surface in degree}

\item{moduleAzi}{Direction the surface is facing}

\item{seaLevel}{Location height above sea level in kilometers}

\item{useBin}{Print solar energy for every day in year if TRUE}
}
\value{
Combined vector of Solar Energy for one year (kWh/m^2/year) and every day (kWh/m^2/day) in that year
}
\description{
Calculate Solar Energy for One Year
}
\examples{
solarOneYear(2015,51.56,15.32,0,30,180,0.145)
}

