% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{SS}
\alias{SS}
\title{Sunrise and Sunset}
\usage{
SS(latitude, DA, TC)
}
\arguments{
\item{latitude}{Location's latitude in degree}

\item{DA}{Declination Angle in degree}

\item{TC}{Time Correction in minutes}
}
\value{
Sunrise and sunset time in hours
}
\description{
Calculate the sunrise and sunset time.
}
\examples{
SS(51.53,2.936844,1)
}

