% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{SunOnTiltedSurface}
\alias{SunOnTiltedSurface}
\title{Solar Statistic Calculation}
\usage{
SunOnTiltedSurface(year, month, day, hour, minute, latitude, longtitude, GMT,
  titlAngle, moduleAzi, seaLevel)
}
\arguments{
\item{year}{Year number}

\item{month}{Month number}

\item{day}{Day number}

\item{hour}{Hour number}

\item{minute}{Minute number}

\item{latitude}{Location's latitude in degree}

\item{longtitude}{Location's longtitude in degree}

\item{GMT}{Location's GMT offset}

\item{titlAngle}{Tilt Angle of the surface in degree}

\item{moduleAzi}{Direction the surface is facing}

\item{seaLevel}{Location height above sea level in kilometers}
}
\value{
Direct beam solar radiation, Solar Radiation on Tilted Surface, Air Mass, Sun Azimuth Angle, Sun Zenith Angle,
Sunrise Time, Sunset Time
}
\description{
Calculate solar energy statistics on tilted surface for a specific date and time.
}
\examples{
SunOnTiltedSurface(2015,6,22,10,0,51.56,15.32,1,30,180,0.145)
}

