% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{TC}
\alias{TC}
\title{Time Correction Factor}
\usage{
TC(longtitude, LSTM, EOT)
}
\arguments{
\item{longtitude}{Location's longtitude in degree}

\item{LSTM}{Local Standard Time Meridian in degree}

\item{EOT}{Equation of Time in minutes}
}
\value{
Time Correction in minutes
}
\description{
The net Time Correction Factor (in minutes) accounts for 
the variation of the Local Solar Time (LST) within a given time zone
}
\examples{
TC(15.5,150,4.753871)
}

