% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suncalculator.R
\name{ZA}
\alias{ZA}
\title{Zenith Angle}
\usage{
ZA(EA)
}
\arguments{
\item{EA}{Elevation Angle in degree}
}
\value{
Zenith Angle in degree
}
\description{
The zenith angle is the angle between the sun and the vertical.
}
\examples{
ZA(30)
}

